Skip locked deployments when analyzing state in health monitor
Health monitor was previously unaware of when a deploy was happening,
causing it to generate alerts unnecessarily.

This change allows health monitor to consider if the deployment is
locked, and it will now skip analyzing agents and instances from locked
deployments. Health monitor should no longer generate duplicate scan and
fix tasks, or incorrect scan and fix tasks during deploys.

[finishes #167502252](https://www.pivotaltracker.com/story/show/167502252)

Co-authored-by: Belinda Liu <bliu@pivotal.io>