Add `initial_deploy_az_update_strategy` to update configuration
This change adds the `initial_deploy_az_update_strategy` field to the
update block. The valid values are `serial` and `parallel`. When
`parallel` is specified, all instances in the instance group will be
updated in parallel on the first deploy regardless of availability zone.

This is in contrast to the default behavior in which BOSH will only
update instances in an instance group one availability zone at a time.

This feature is an opt-in feature as updating instances in parallel
across availability zones is not guaranteed to be safe depending on the
software that BOSH is deploying.

[finishes #168067218](https://www.pivotaltracker.com/story/show/168067218)

Co-authored-by: James Myers <jmyers@pivotal.io>