Bring back chruby in unit and integration tests
Previously the chruby command was changed/commented out since we thought
it was removed from the mysql-5.7 docker image. Chruby is in fact
installed through our install-ruby script, and as it is still necessary
for older LTS versions, we are reverting back to the original command in
our testing scripts.

[#168653211](https://www.pivotaltracker.com/story/show/168653211)
[#168222310](https://www.pivotaltracker.com/story/show/168222310)

Co-authored-by: Rebecca Putinski <rputinski@pivotal.io>